a68: fix type of flex.sub_offset in struct encoded_mode
The type of flex.sub_offset in struct encoded_mode shall of course be
uint64_t rather than uint8_t.  This was triggering corrupted exports
sections once a certain amount of modes were reached.


gcc/algol68/ChangeLog

	* a68-imports.cc (struct encoded_mode): Type of flex.sub_offset
	shall be uint64_t.

+++ gcc/algol68/a68-imports.cc
@@ -650,7 +650,7 @@ struct encoded_mode
 
     struct
     {
-      uint8_t sub_offset;
+      uint64_t sub_offset;
     } flex;
 
     struct
